chore(plugins): sync FixedShipping with upstream shopplugins/FixedShipping.git
Adopt upstream changes:
- Boot.php: cast 'value' setting to (float) for type safety
- Public/images/logo.png: update to latest icon (blue-grey theme)
- config.json: adopt 4-space JSON indentation to match upstream style

+++ plugins/FixedShipping/Boot.php
@@ -50,7 +50,7 @@ class Boot extends BaseBoot
     {
         $subtotal      = $entity->getSubtotal();
         $shippingType  = plugin_setting('fixed_shipping', 'type', 'fixed');
-        $shippingValue = plugin_setting('fixed_shipping', 'value', 0);
+        $shippingValue = (float) plugin_setting('fixed_shipping', 'value', 0);
         if ($shippingType == 'fixed') {
             return $shippingValue;
         } elseif ($shippingType == 'percent') {
